moving to es6 modules
this breaks all the inline scripts, like onclick attributes on buttons.
