+++ docs/index.html
@@ -19,7 +19,8 @@
     <script src="js/libs/countup.min.js"></script>
     <!-- page -->
     <link type="text/css" rel="stylesheet" href="css/index.css?c=3"/>
-    <script src="js/main.js?c=3"></script>
+    <script type="module" src="js/main.js?c=3"></script>
+    <script nomodule src="js/main.js?c=3"></script>
 
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
 </head>
