+++ docs/wall.html
@@ -66,12 +66,12 @@
                                                target="_blank"></a></li>
     </ul>
 
-    <div class="container">
+    <div class="container" >
       <div class="section center-align">
         <h4 class="no-margin-top" >Top Traders wall</h4>
         <p>Select one of the market seasons below to view it's top 50 traders.</p>
         <a href="#info1" class="waves-effect waves-light btn grey darken-3">beta</a>
-        <a href="#info1" class="waves-effect waves-light btn grey darken-3 disabled"><i class="material-icons right">expand_more</i>select</a>
+    <!--<a href="#info1" class="waves-effect waves-light btn grey darken-3 disabled"><i class="material-icons right">expand_more</i>select</a> -->
         <a href="#info1" class="waves-effect waves-light btn grey darken-3 disabled">current</a>
         
       <div class="row top3-cards">
@@ -206,6 +206,8 @@
             </tbody>
         </table>        
         </div>
+        <br>
+        <a href="#" class="btn-floating btn-large waves-effect waves-light grey darken-3"><i class="material-icons">expand_less</i></a>
       </div>
     </div>
